requirements.txt
numpy==1.15.4
pandas==0.23.4
psutil==5.4.8
PyYAML==3.13
gin-config==0.1.2
tensorflow==1.12.0
keras==2.2.4
sklearn==0.0
scikit-learn==0.20.1
google-cloud-storage==1.14.0